+++ NEWS.md
@@ -1,3 +1,8 @@
+# errors devel
+
+- Add option `decimals` to `format()` method to add support for uncertainty with
+  decimals in the `parenthesis` notation (@alusiani #60, #61 addressing #47).
+
 # errors 0.4.2
 
 - Add support for PDG rounding rules (@davidchall #59 addressing #45).